@@ -4,24 +4,35 @@ version: 2.1.0
 homepage: https://github.com/benjamindean/flutter_vibration
 
 environment:
-  sdk: ">=2.12.1 <4.0.0"
-  flutter: ">=1.20.0"
+  sdk: ^3.6.0
+  flutter: ">=3.3.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  # TODO: Change pub.dev's version when vibration_web is published
-  # vibration_web: ^1.6.6
+  plugin_platform_interface: ^2.0.2
   vibration_platform_interface: ^0.0.2
 
-# dependency_overrides:
-#   vibration_platform_interface:
-#     path: ../vibration_platform_interface
 dev_dependencies:
   flutter_test:
     sdk: flutter
+  flutter_lints: ^5.0.0
 
+# For information on the generic Dart part of this file, see the
+# following page: https://dart.dev/tools/pub/pubspec
+
+# The following section is specific to Flutter packages.
 flutter:
+  # This section identifies this Flutter project as a plugin project.
+  # The 'pluginClass' specifies the class (in Java, Kotlin, Swift, Objective-C, etc.)
+  # which should be registered in the plugin registry. This is required for
+  # using method channels.
+  # The Android 'package' specifies package in which the registered class is.
+  # This is required for using method channels on Android.
+  # The 'ffiPlugin' specifies that native code should be built and bundled.
+  # This is required for using `dart:ffi`.
+  # All these are used by the tooling to maintain consistency when
+  # adding or updating assets for this project.
   plugin:
     platforms:
       android:
@@ -29,5 +40,34 @@ flutter:
         pluginClass: VibrationPlugin
       ios:
         pluginClass: VibrationPlugin
-      # web:
-      #   default_package: vibration_web
+
+  # To add assets to your plugin package, add an assets section, like this:
+  # assets:
+  #   - images/a_dot_burr.jpeg
+  #   - images/a_dot_ham.jpeg
+  #
+  # For details regarding assets in packages, see
+  # https://flutter.dev/to/asset-from-package
+  #
+  # An image asset can refer to one or more resolution-specific "variants", see
+  # https://flutter.dev/to/resolution-aware-images
+
+  # To add custom fonts to your plugin package, add a fonts section here,
+  # in this "flutter" section. Each entry in this list should have a
+  # "family" key with the font family name, and a "fonts" key with a
+  # list giving the asset and other descriptors for the font. For
+  # example:
+  # fonts:
+  #   - family: Schyler
+  #     fonts:
+  #       - asset: fonts/Schyler-Regular.ttf
+  #       - asset: fonts/Schyler-Italic.ttf
+  #         style: italic
+  #   - family: Trajan Pro
+  #     fonts:
+  #       - asset: fonts/TrajanPro.ttf
+  #       - asset: fonts/TrajanPro_Bold.ttf
+  #         weight: 700
+  #
+  # For details regarding fonts in packages, see
+  # https://flutter.dev/to/font-from-package
